backend/requirements.txt
# Backend dependencies — MVP 1
# Python 3.11+

# --- Web framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# --- Database ---
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
greenlet==3.1.1

# --- Settings / validation ---
pydantic==2.10.4
pydantic-settings==2.7.1

# --- Task queue ---
celery[redis]==5.4.0
redis==5.2.1

# --- MinIO / S3 ---
minio==7.2.12

# --- PDF ---
PyMuPDF==1.25.1

# --- DOCX ---
python-docx==1.1.2

# --- HTTP client ---
httpx==0.27.2

# --- OpenAI ---
openai==1.51.0

# --- Utils ---
Pillow==11.1.0
